bugfix: dont assign `type/topic` to generated topic page...
... this is because it ends up *somehow* confusing the topic collection so that topics with a written taxonomy page don't exclude the autogenerated page, therefore causing both to exist and 11ty to crash with a duplicate permalink exception.

When attaching a debugger it seemed that 11ty was processing topic.njk *while* the topics collection was being built. Which is odd because topic.njk depends upon that collection.

After spending more than an hour delving into this issue I decided to simply remove the tag from the autogenerated topic page's and set page_type instead. This is *only* to be used for the PageFind index anyway; a lot of effort just so a box says 'topic' instead of 'index' lol.